Remove DocumentCreatedEvent and move behaviour to ChildCreatedEvent
